refactored node/cert creation
